requirements.txt
absl-py==2.1.0
accelerate==1.1.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiohttp-cors==0.8.0
aiosignal==1.3.1
airportsdata==20250224
annotated-types==0.7.0
anyio==4.7.0
astor==0.8.1
async-timeout==5.0.1
attrs==24.2.0
audioread==3.0.1
av==14.0.1
bitsandbytes==0.45.0
blake3==1.0.4
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorful==0.5.6
compressed-tensors==0.9.1
contourpy==1.3.1
cupy-cuda12x==13.4.0
cycler==0.12.1
datasets==3.1.0
decorator==5.1.1
depyf==0.18.0
diffusers==0.31.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
einops==0.8.0
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-cli==0.0.7
fastrlock==0.8.3
ffmpy==0.4.0
fonttools==4.55.2
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2024.9.0
gguf==0.10.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.2
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
gradio==5.8.0
gradio_client==1.5.1
grpcio==1.68.1
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.0
huggingface-hub==0.26.3
identify==2.6.3
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
interegular==0.3.3
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lark==1.2.2
lazy_loader==0.4
librosa==0.10.2.post1
linkify-it-py==2.0.3
llava==0.0.1.dev0
llvmlite==0.43.0
lm-format-enforcer==0.10.11
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.3
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.16.0
mistral_common==1.5.1
mkl-service==2.4.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
ninja==1.11.1.2
nodeenv==1.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.57.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
optree==0.13.1
orjson==3.10.12
outcome==1.3.0.post0
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
pandas==2.2.3
partial-json-parser==0.2.1.1.post4
peft==0.13.2
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
pre_commit==4.0.1
prettytable==3.12.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.1
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.1
psutil==6.1.0
py-cpuinfo==9.0.0
py-spy==0.4.0
pyairports==2.1.1
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycountry==24.6.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
pyecharts==2.0.7
Pygments==2.18.0
pyparsing==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
pyzmq==26.2.0
qwen-vl-utils==0.0.8
ray==2.40.0
referencing==0.35.1
regex==2024.11.6
requests 
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.22.3
rsa==4.9
ruff==0.8.2
safehttpx==0.1.6
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.27.1
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.19.1
setproctitle==1.3.4
shellingham==1.5.4
shortuuid==1.0.13
simplejson==3.19.3
six==1.17.0
smart-open==7.1.0
smmap==5.0.1
snapshot-selenium==0.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.5.0.post1
starlette==0.41.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
textual==2.1.2
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.13
tokenizers==0.21.0
tomli==2.2.1
tomlkit==0.13.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.50.0
trio==0.28.0
trio-websocket==0.11.1
triton==3.1.0
typer==0.15.1
types-requests==2.32.0.20241016
tzdata==2024.2
uc-micro-py==1.0.3
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
virtualenv==20.28.0
vllm==0.7.3
wandb==0.19.0
watchfiles==1.0.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xformers==0.0.28.post3
xgrammar==0.1.11
xxhash==3.5.0
yarl==1.18.3
yt-dlp==2024.12.23
zipp==3.21.0
